+++ migrations/095_course_instances__ps_linked__add.sql
@@ -1,0 +1,1 @@
+ALTER TABLE course_instances ADD COLUMN ps_linked BOOLEAN NOT NULL DEFAULT TRUE;